Bug 1117468 - Fix Android gcc warnings in js/src/irregxp and js/src/jit. r=jandem
UltraBlame original commit: b8ade3d51d5ccdc93113187c231831db5a27ebaf

+++ js/src/irregexp/RegExpEngine.cpp
@@ -82,7 +82,7 @@ static const int kSurrogateRangeCount = ArrayLength(kSurrogateRanges);
 static const int kLineTerminatorRanges[] = { 0x000A, 0x000B, 0x000D, 0x000E,
     0x2028, 0x202A, 0x10000 };
 static const int kLineTerminatorRangeCount = ArrayLength(kLineTerminatorRanges);
-static const unsigned kMaxOneByteCharCode = 0xff;
+static const int kMaxOneByteCharCode = 0xff;
 static const int kMaxUtf16CodeUnit = 0xffff;
 
 static char16_t
@@ -3180,7 +3180,7 @@ SplitSearchSpace(RangeBoundaryVector &ranges,
     
     
     
-    if (*border - 1 > (int) kMaxOneByteCharCode &&  
+    if (*border - 1 > kMaxOneByteCharCode &&  
         end_index - start_index > (*new_start_index - start_index) * 2 &&
         last - first > kSize * 2 &&
         binary_chop_index > *new_start_index &&
